edited database

+++ UFAFAC/Sources/DataStructure/StringsHolder.cpp
@@ -4,26 +4,26 @@
 
 using namespace DataStructure;
 
-const std::wstring& StringsHolder::GetString(u64 index)
+const std::wstring& StringsHolder::GetString(u32 index)
 {
 	assert(index < strings.size());
 	strings[index].refCount++;
 	return strings[index].entry;
 }
 
-void StringsHolder::IncrementRef(u64 index)
+void StringsHolder::IncrementRef(u32 index)
 {
 	assert(index < strings.size());
 	strings[index].refCount++;
 }
 
-void StringsHolder::DecrementRef(u64 index)
+void StringsHolder::DecrementRef(u32 index)
 {
 	assert(index < strings.size());
 	strings[index].refCount--;
 }
 
-void StringsHolder::ReleaseString(u64 index)
+void StringsHolder::ReleaseString(u32 index)
 {
 	assert(index < strings.size() && strings[index].refCount);
 	strings[index].refCount--;
@@ -34,26 +34,26 @@ void StringsHolder::ReleaseString(u64 index)
 	}
 }
 
-u64 StringsHolder::FindOrCreateString(const std::wstring& str)
+u32 StringsHolder::FindOrCreateString(const std::wstring& str)
 {
-	u64 res = FindString(str);
+	u32 res = FindString(str);
 	if (res != -1) return res;
 	if (!availableSlots.empty())
 	{
-		u64 slot = availableSlots.front();
+		u32 slot = availableSlots.front();
 		strings[slot] = StringInfo(str);
-		unorderedStrings.insert(std::pair<std::wstring, u64>(str, slot));
+		unorderedStrings.insert(std::pair<std::wstring, u32>(str, slot));
 		std::copy(availableSlots.data() + 1, availableSlots.data() + availableSlots.size(), availableSlots.data());
 		availableSlots.pop_back();
 		return slot;
 	}
 	StringInfo info = StringInfo(str);
 	strings.push_back(str);
-	unorderedStrings.insert(std::pair<std::wstring, u64>(str, strings.size() - 1));
+	unorderedStrings.insert(std::pair<std::wstring, u32>(str, strings.size() - 1));
 	return strings.size() - 1;
 }
 
-u64 StringsHolder::FindString(const std::wstring& str)
+u32 StringsHolder::FindString(const std::wstring& str)
 {
 	auto t = unorderedStrings.find(str);
 	if (t != unorderedStrings.end())
